Restored old motor library behaviour for backwards compatibility.
Made modifications to the motor library so that the old style pin: x version
will work for non-directional motors. Inclusion of a motor type now used
internally to specify the type of motor being used.

Reinstated the old motor demo and tested that it works with the new library and
moved the newer demo with directions to its own demo.