Add servo.fps property
This needs more hardware testing before merging.

Adds servo.fps (Frames Per Second)

Passing rate param to Servo.to() will override fps for that call

When rate exceeds the number of steps possible for the given PWM
resolution and distance traveled we override both fps and rate and use
a calculated value.

I am not checking to see if the time allows for the distance travelled,
but that could be something to consider.