Support for inverting speed when direction changes
During JSConf I learned that some motor controllers need the PWM value
inversed when the direction is switched. Most commercial motor
controllers abstract this need away but basic H-Bridge controllers do
not.

This adds an optional property to the motor ops object:

```
var motor = new five.Motor({
  pins: [3, 12],
  invertSpeed: true
});
```
When invertSpeed is true and the dir pin is set to high, the pwm pin
value will be set to 255 - speed

Thanks to @JonHylands for his explanations. Any mistakes herein are
mine, not his.