Timing is everything
In the playTune test there was just 4ms of wiggle room for the tune to
finish. In Travis this failed about half the time (only 3 calls were
made on freqSpy). I added 20ms of wiggle room to the timeout.

Also in toneStopsAfterTime the test was inermittently firing before the
tone had finished playing. I added 10 ms of wiggle room there.