RGB.on() and .off() saves color state
Led.RGB.on() and Led.RGB.off() weren’t saving color state.

It was also treating the RGB as individual LEDs and calling Led.on()
was causing unexpected results because there isn’t the concept of
individual brightness on the RGB pins. (It just produced unexpected
color combinations rather than “dimming” them.)

This simplified implementation will need to be revisited when/if #375
is addressed, but at least it’s consistent and intuitive for now.
Led.RGB.on() is essentially the same as Led.RGB.color(“#ffffff”) with
this update.

Color also wasn’t being checked for in the shape tests for RGB (both
varieties)

Fixes #445