Feature implementation and bug fixes on RBG.
Adds isOn property for RGB that looks to see if ANY of the colors are
on. When calling .on(), it leaves the led as-is if it’s already on.

Adds chaining capability to .on() and .off()

Supports saving previous color as requested in #375

Updates toggle/strobe/stop functions to work with RGB.

Removes functions not applicable to RBG
(fade, fadeIn, fadeOut, pulse, brightness)
Note, we could potentially implement these later with appropriate calculations.