Fixes to servo
Addresses #666

1) Range was being constrained after offset was applied which was
wrong. User should not have to think in two different coordinate spaces.
2) Invert was based on global coordinate range which led to incorrect
values being passed to the servo. Invert has to be based on the
physical range of the servo in local coordinate space. We need to
maintain to ranges, the global coordinate range for ease of use and the
physical device range of the servo. We are assuming 180° servos for
now. We will have to add in 45° and 270° servos as Devices.
3) Added tests