Fix default pwmRange for PCA9685
These new values were tested with an eMax and a Hitec. Our new
Servo.sweep() with easing really made it clear that we were hitting the
top of the range too early and through experimentation I found we were
missing some of the low end range.

Revert "Add to .tesselignore"

This reverts commit 9f0bd313619e02702085b16a619709302d43dc01.