Fix Piezo.prototype.note() (#1328)
* Fix Piezo.prototype.note()

The `note()` method was passing a frequency (Hz) to the `tone()` method, which is expecting to receive a duty cycle (μs), leading to the wrong note being played.  The solution is for `note()` to call `frequency()` instead, which will do the appropriate conversion.

Since it took me a while to understand the relationships between `tone`, `frequency` and `note` I've added some documentation to those methods to clear things up for the next person.  I've also made some unit tests more aggressive about checking that values are converted appropriately as they're passed around, and commented note-frequency-duty equivalencies to help the test read nicer.

* Fix lint error

* Address review feedback

- Test that `note()` is case-insensitive
- Better description of default octave
- One-line Piezo.prototype.note
- s/frequncy/frequency
- Note tests demonstrate the difference between note letter and octave number better