Customize package.json and README for Code.org
Adds a note near the top of the README that this is our fork, and linking to the original repo.  Hides badges not relevant to our repo, updates Travis badge to point at our builds.

Adjusts package name and repo links in package.json for better metadata on npmjs.org.