Animation uses fallback timer immediately
The 'temporal' library being used for animation in johnny-five, while great for robotics, is too CPU-intensive for use in Maker Lab.  There's already a mechanism for falling back on a setInterval-based solution for animations that run longer than 5 seconds.  By tuning that fallback timeout down to zero we can use the setInterval fallback immediately and avoid the performance impact of temporal.

Animation is the only place in johnny-five that temporal is used, so we don't have to do this anywhere else.

If upstream ever makes this fallback officially configurable, we won't need this customization anymore.