Serial: throttle serial list attempts, add forced exit
When listing available serial connections, a maximum number of 10
attempts every 400 milliseconds are allowed, then the `board.fail`
method is called.
When disconnecting on SIGINT, add a 1000 millisecond timeout before
forcing the exit. This has been added because if one disconnects a
succesfully connected board the process could never exit on SIGINT.