Repl: fix multi board disconnection
When connecting multiple boards, an exception is thrown when exiting the
process
because `state.board.io.pending` is not defined. This commit fixes the
problem by checking the `length` property of `state.board` and looping
through the board list to check the `board.io.pending` property.
Futhermore, it add a timeout of 1000 milliseconds to force the process
exit in case the `board.io.pending` property never becomes `false` (it
has happened disconnetting a succesfully connected board).