GitHub based website build system
This Hosts and Builds johnny-five.io with github pages and actions

This would replace our current setup with one that’s completely GitHub based. It uses GitHub pages with custom domains and https enabled for hosting, and GitHub actions to build the site automatically whenever the code or wiki changes.

On a commit to master or update to the wiki
- clone nodebots/johnny-five.io
- npm ci && npm build
- force push the `public` folder to the `gh-pages` branch of rwaldron/johnny-five

The build can be triggered manually be rerunning the last job in the actions tab on GitHub.

This requires the beta (public release in November) of GitHub actions. If you aren’t already in it you can request access. (I requested it for the `nodebots` org btw.) https://github.com/features/actions

You can see a demo (from a fork) here: http://reconbot.github.io/johnny-five/
(Assets aren’t configured to work without a custom domain which I didn’t setup for this demo)

You can see all the logs in the actions tab of the repo.
https://github.com/reconbot/johnny-five/actions