Suspendable: adds class that provides a "pause" and "resume" method
We can't safely or effectively destroy objects that represent hardware or hardware states, but we need a way to temporarily prevent sensors from emitting data events. This class will move the contents of _events into a side table weakmap when pause() is called and then restore _events when resume() is called.